backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pillow==10.4.0
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
